requirements.txt
matplotlib==3.10.7
numpy==2.3.5
pandas==2.3.3
Pillow==12.0.0
scipy==1.16.3
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
